Added a Pause Menu
In a new scene, added a pause menu.

To do this, I created a new script (PauseMenu), as well as a few animations. There is now an 'Animations' Folder in our project.

Upon hitting Escape, game pauses and a UI element slides in with an animation. Upon hitting escape again, game unpauses.