Character Status Menu
On the pause menu, implemented the status screens for the characters. There is flexible padding on the menu, so whether or not there is one, two, or three current party members, they space themselves out on the menu display.

Each player info graphic in the menu comes with several status bars for the character, which is used to convey to the player information about said character.