Screen Space issues with the menu
My suspicions about the world-space versus the screen space seemed to be true. I don't understand why there were issues, however. All of the animated components of the UI are children of the main canvas, which occupies the same space as the Camera view. I am unsure why issues arose.