Started Inventory Programming
Inventory is managed by the InventoryMenuLogic, which is a script attached to a Gameobject which encompases the horizontal group that displays inventory status.

The inventory is kept in a List of type InventoryItem, which is a specified nested class of InventoryMenuLogic.

My plans for the future is to make it so the inventory menu calls the list and displays the correct items, each item is only displayed once (ie: player has 5 apples but there is only one apple icon with an indicator that the player has 5), and the menu for selecting an inventory item to use has a dynamic scroll element like a rotunda.

I'm not a stellar programmer but I think this is along the lines of what we need to do here.